move requestResponse and requestStream to separate files
