+++ packages/gateway/src/Gateway.ts
@@ -1,28 +1,25 @@
-import { Observable } from 'rxjs';
-// @ts-ignore
 import { RSocketServer } from 'rsocket-core';
-// @ts-ignore
 import RSocketWebSocketServer from 'rsocket-websocket-server';
-// @ts-ignore
-import { Single, Flowable } from 'rsocket-flowable';
 import { Api } from '@scalecube/scalecube-microservice';
+import { requestResponse } from './requestResponse';
+import { requestStream } from './requestStream';
 
 export class Gateway implements Api.Gateway {
   private port: number;
-  private server: any;
-  private transport: any;
+  private server;
+  private transport;
   constructor(options) {
     const { port } = options;
     this.port = port;
     this.transport = new RSocketWebSocketServer({ port });
   }
-  start(opts: any) {
+  start(opts) {
     const { serviceCall } = opts;
     this.server = new RSocketServer({
-      getRequestHandler: (socket: any) => {
+      getRequestHandler: (socket) => {
         return {
-          requestResponse: (payload: any) => requestResponse(payload, serviceCall),
-          requestStream: (payload: any) => requestStream(payload, serviceCall),
+          requestResponse: (payload) => requestResponse(payload, serviceCall),
+          requestStream: (payload) => requestStream(payload, serviceCall),
         };
       },
       transport: this.transport,
@@ -35,41 +32,3 @@ export class Gateway implements Api.Gateway {
     console.log('Gateway stopped');
   }
 }
-
-const requestResponse = (payload: any, serviceCall: any) => {
-  const { data, metadata } = payload;
-  console.log('Request:', data, metadata);
-  return new Single((subscriber: any) => {
-    subscriber.onSubscribe();
-    serviceCall({
-      message: JSON.parse(data),
-      asyncModel: 'requestResponse',
-      // includeMessage: true,
-    })
-      .then((resp: any) => {
-        console.log('RESP', resp);
-        subscriber.onComplete({ data: JSON.stringify(resp), metadata: 'my_response_metadata' });
-      })
-      .catch((err: any) => {
-        subscriber.onError(err);
-      });
-  });
-};
-const requestStream = (payload: any, serviceCall: any) => {
-  console.log('request payload: ', payload);
-  const { data, metadata } = payload;
-  return new Flowable((subscriber: any) => {
-    subscriber.onSubscribe();
-    const message = JSON.parse(data);
-    (serviceCall({
-      message,
-      asyncModel: 'requestStream',
-    }) as Observable<any>).subscribe(
-      (response: any) => {
-        subscriber.onNext({ data: JSON.stringify(response), metadata: '' });
-      },
-      (error: any) => subscriber.onError(error),
-      () => subscriber.onComplete()
-    );
-  });
-};
